core/future: Add ability to walk continuation chains
This patch introduces task::waiting_task() which allows one to walk
the chain of tasks which are currently blocked on each other. The
invocation gives a pointer to the task which is waiting for the
subject of the call to complete execution.

By generalization, we also have now supporting methods of similar
meaning on different kinds of objects as well: promise::waiting_task()
and thread::waiting_task().

This ability to walk the continuation chain will be used to implement
better backtracing, one which gives more context. This will be
explained in detail in later commit.