core/task, util/backtrace: Collect backtraces across continuation chains
Previous saved_backtrace is now called simple_backtrace.

tasktrace is a simple_backtrace of the currently running task plus a
collection of backtraces corresponding to the continuation chain
waiting for the current task to complete.

Containuation chain is a sequence of tasks such that the next task
runs when the previous one resolves and the next task depends on the
result of the previous task.

Each task is represented in the tasktrace by the name of the task type.

Later we will allow configuring seastar at compile-time to
collect full backtrace at preemption point and save it in the blocked
task.

For example, when a thread is waiting for I/O operation to complete,
the chain will contain the continuation which handles ready I/O, then
a continuation which wakes the thread, then the continuation which
waits for the thread to finish.

This gives much more context in the backtrace, similar to the one we
would get in a synchronous programming model.

Another use case is heap profiles. Currently measuring memory
allocations for one large operation is hard if it starts many
allocating continuations. Tasks for those continuations will be spread
in the profile and not correlated with the grand
operation. Backtracing across tasks solves that because different
paths are joined by a common base.